This is an early implementation of a vertical waymarking algorithm. The idea is to eliminate Use::Val, thus shrinking struct Use to 8 bytes.
Of course this needs some recalculation logic
to recover the Value* by walking the def/use
chain. Whether the computing overhead outweighs
the space savings is an open question at this
point.

There are some tests at the end, so far all
pass.

Please Note: this is a private branch, this file
would break the llvm build, I'll fix this later.

llvm-svn: 56160