Pull 65497 and 65498 into Dib:
If an available register falls through to a succ block, unset the last
kill. Sorry, it's impossible to reduce a sensible test case. It basically
requires the moon and stars to align in order to cause a failure.

The last commit was overly conservative. It's ok to reuse value that's already
marked livein.

llvm-svn: 65515