--- Merging r70270 into '.': U    test/DebugInfo/forwardDecl.ll U    test/DebugInfo/2008-11-05-InlinedFuncStart.ll U    test/DebugInfo/deaddebuglabel.ll U    test/DebugInfo/2009-01-30-Method.ll U    test/CodeGen/PowerPC/cr_spilling.ll U    test/CodeGen/X86/2008-04-15-LiveVariableBug.ll U    test/CodeGen/X86/add-with-overflow.ll ...
Massive check in. This changes the "-fast" flag to "-O#" in llc. If you want to
use the old behavior, the flag is -O0. This change allows for finer-grained
control over which optimizations are run at different -O levels.

Most of this work was pretty mechanical. The majority of the fixes came from
verifying that a "fast" variable wasn't used anymore. The JIT still uses a
"Fast" flag. I'm not 100% sure if it's necessary to change it there...

llvm-svn: 70272